fix(media): keep fenced-MEDIA latch off public reply-payload SDK
Move createDirectAcceptedFencedMediaWarnLatch to the private-local
channel-outbound-fenced-media runtime seam so bundled channels can share
the accepted-send diagnostic without expanding the public Plugin SDK
reply-payload contract (#41966 / ClawSweeper P1).
