+++ scripts/lib/plugin-sdk-private-local-only-subpaths.json
@@ -21,6 +21,7 @@
   "channel-config-writes",
   "channel-contract-testing",
   "channel-mention-gating",
+  "channel-outbound-fenced-media-runtime",
   "channel-route",
   "channel-secret-tts-runtime",
   "channel-target-testing",
@@ -30,7 +31,6 @@
   "cli-backend",
   "cli-runtime",
   "codex-mcp-projection",
-  "native-hook-relay-runtime",
   "codex-session-transcript-runtime",
   "command-status-runtime",
   "command-surface",
@@ -69,8 +69,8 @@
   "memory-core-host-engine-curated",
   "memory-core-host-engine-embeddings",
   "memory-core-host-engine-fs",
-  "memory-core-host-engine-sessions",
   "memory-core-host-engine-schema",
+  "memory-core-host-engine-sessions",
   "memory-core-host-engine-storage",
   "memory-core-host-runtime-cli",
   "memory-core-host-runtime-core",
@@ -85,6 +85,7 @@
   "migration-runtime",
   "model-ref-parse",
   "music-generation",
+  "native-hook-relay-runtime",
   "node-host",
   "number-runtime",
   "outbound-echo-runtime",
