+++ src/plugin-sdk/channel-outbound-fenced-media-runtime.ts
@@ -3,6 +3,7 @@
  *
  * Not a public third-party Plugin SDK contract. Bundled Telegram/Slack/Signal
  * direct paths share the accepted-delivery warn helper without expanding
- * `openclaw/plugin-sdk/channel-outbound`.
+ * `openclaw/plugin-sdk/channel-outbound` or public `reply-payload`.
  */
-export { warnFencedMediaSkipsForAcceptedOutboundDelivery } from "../infra/outbound/fenced-media-skip-warn.js";
+export { warnFencedMediaSkipsForAcceptedOutboundDelivery } from "./channel-outbound-fenced-media-warn.js";
+export { createDirectAcceptedFencedMediaWarnLatch } from "./channel-outbound-fenced-media-latch.js";
