+++ src/plugin-sdk/channel-outbound-fenced-media-warn.ts
@@ -1,0 +1,1 @@
+export { warnFencedMediaSkipsForAcceptedOutboundDelivery } from "../infra/outbound/fenced-media-skip-warn.js";
