+++ src/plugin-sdk/reply-payload.fenced-media.test.ts
@@ -2,16 +2,11 @@
 import { describe, expect, it, vi } from "vitest";
 
 const warnFencedMediaSkipsForAcceptedOutboundDelivery = vi.hoisted(() => vi.fn());
-vi.mock("./channel-outbound-fenced-media-runtime.js", async (importOriginal) => {
-  const actual =
-    await importOriginal<typeof import("./channel-outbound-fenced-media-runtime.js")>();
-  return {
-    ...actual,
-    warnFencedMediaSkipsForAcceptedOutboundDelivery: (
-      ...args: Parameters<typeof warnFencedMediaSkipsForAcceptedOutboundDelivery>
-    ) => warnFencedMediaSkipsForAcceptedOutboundDelivery(...args),
-  };
-});
+vi.mock("./channel-outbound-fenced-media-warn.js", () => ({
+  warnFencedMediaSkipsForAcceptedOutboundDelivery: (
+    ...args: Parameters<typeof warnFencedMediaSkipsForAcceptedOutboundDelivery>
+  ) => warnFencedMediaSkipsForAcceptedOutboundDelivery(...args),
+}));
 
 import {
   deliverFormattedTextWithAttachments,
