+++ src/plugin-sdk/reply-payload.ts
@@ -2,7 +2,6 @@
 import { normalizeLowercaseStringOrEmpty } from "../../packages/normalization-core/src/string-coerce.js";
 import type { ReplyPayload as InternalReplyPayload } from "../auto-reply/reply-payload.js";
 import type { ChannelOutboundAdapter } from "../channels/plugins/outbound.types.js";
-import { createOutboundPayloadPlan } from "../infra/outbound/payloads.js";
 import { normalizeOutboundReplyPayloadCore as normalizeCoreOutboundReplyPayload } from "../infra/outbound/reply-payload-normalize.js";
 import {
   countOutboundMedia,
@@ -13,7 +12,7 @@ import {
 } from "../infra/outbound/reply-payload-parts.js";
 import { createReplyToFanout } from "../infra/outbound/reply-policy.js";
 import { hasReplyPayloadContent } from "../interactive/payload.js";
-import { warnFencedMediaSkipsForAcceptedOutboundDelivery } from "./channel-outbound-fenced-media-runtime.js";
+import { createDirectAcceptedFencedMediaWarnLatch } from "./channel-outbound-fenced-media-runtime.js";
 
 export type { MediaPayloadInput } from "../channels/plugins/media-payload.js";
 /** @deprecated Inbound contexts use `media`; outbound replies use `ReplyPayload.mediaUrl(s)`. */
@@ -168,61 +167,6 @@ export function resolveTextChunksWithFallback(text: string, chunks: readonly str
     return [];
   }
   return [text];
-}
-
-/** Latch fenced-MEDIA diagnostics to accepted visible direct-delivery text (#41966). */
-export function createDirectAcceptedFencedMediaWarnLatch(params: {
-  payload: object;
-  cfg?: unknown;
-  surface?: string;
-}) {
-  // SAFETY: channel deliver path feeds already-normalized ReplyPayload-shaped objects into the plan builder
-  const planEntry = createOutboundPayloadPlan([params.payload as never], {
-    cfg: params.cfg as never, // SAFETY: OpenClawConfig structural match across plugin-sdk/core boundary
-    surface: params.surface,
-  })[0];
-  if (!planEntry?.mediaTokenSkippedInFence) {
-    return {
-      afterAcceptedVisibleText(_chunk: string) {},
-    };
-  }
-  let warned = false;
-  let acceptedVisibleText = "";
-  const identities = planEntry.fencedSkippedMediaDirectives ?? [];
-  return {
-    afterAcceptedVisibleText(visibleChunk: string) {
-      if (warned) {
-        return;
-      }
-      const chunk = visibleChunk.trim();
-      if (chunk) {
-        acceptedVisibleText = acceptedVisibleText
-          ? `${acceptedVisibleText}\n${visibleChunk}`
-          : visibleChunk;
-      }
-      const retained =
-        identities.length > 0
-          ? identities.some((directive) => {
-              const identity = directive.trim();
-              return (
-                identity.length > 0 &&
-                acceptedVisibleText.split("\n").some((line) => line.trim() === identity)
-              );
-            })
-          : /media:/i.test(acceptedVisibleText);
-      if (!retained) {
-        return;
-      }
-      warned = true;
-      warnFencedMediaSkipsForAcceptedOutboundDelivery([
-        {
-          text: acceptedVisibleText,
-          mediaTokenSkippedInFence: true,
-          fencedSkippedMediaDirectives: identities,
-        },
-      ]);
-    },
-  };
 }
 
 /** Send media-first payloads intact, or chunk text-only payloads through the caller's transport hooks. */
